backend/requirements.txt
fastapi==0.110.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.14
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.25.2
